+++ src/Sentry.AspNetCore/SentryMiddleware.cs
@@ -127,6 +127,9 @@ namespace Sentry.AspNetCore
 
                 void CaptureException(Exception e)
                 {
+                    e.Data[Mechanism.HandledKey] = true;
+                    e.Data[Mechanism.MechanismKey] = "SentryMiddleware";
+
                     var evt = new SentryEvent(e);
 
                     _logger.LogTrace("Sending event '{SentryEvent}' to Sentry.", evt);